+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 ### Bug fixes
 - Fixed NullPointException when no `requires_pictures` key was set in any mod map inside the `.toml` file when mods where imported
+- Fixed #118 - The customized publisher icon was not used when publishers where imported, instead the default icon name was used
 
 ## [v4.8.0](https://github.com/LMH01/MGT2_Mod_Tool/releases/tag/v4.8.0) (Latest Version)